renders js based on community
